- Essentially, treats URLs as just another place to find a package meeting certain requirements,   rather than a special host for only development packages, where version pinning is essentially   offloaded to the revision number of the VCS, if it's used at all. This is a much more general   way of handling non-index-controlled packages; for example, it cleanly supports multiple   Python packages within different directories of the same git repository.
We had a need to dramatically speed up the installation of an internal package managed on GitHub,
which was not only quite large (~100MB working tree), but which required many minutes of Cythoning

Since this somewhat egregiously breaks backward compatibility, it is submitted here only as a
tool for discussion (although it could very well serve as a base commit for future work).
Honestly, I'm not quite sure how to integrate it back in while maintaining compatibility, without
overcomplicating the interface. It's served us quite well as an internal fork, though, since we're
aware of its differences in behavior.

Breaks backward compatibility in the following ways:

- #egg fragments with hyphens will now be parsed (resulting in an error if the parsed version is
  invalid, and a different package name if it isn't)
- like indexed requirements, vcs/file-controlled packages whose conditions are already met will
  not be re-downloaded - for URLS with '#egg=foo', this means that 'foo' will be installed once,
  and only upgraded if --upgrade is passed
- wheels preferred
- if the package cannot be found at the URL for some reason (i.e. _link_package_versions returns
  []), then other sources for it, like the index or --find-links, will be used instead

TODO: tests
TODO: fix breakage with -d and #egg fragment without version